Change value for AWS_*
AWS returns a weird `sts:GetCallerIdentity: IncompleteSignature: 'value' not a valid key=value pair (missing equal-sign)` when the value of the ENV vars is not a simple one word value. This is harder to debug. Switching to `no-value-here` helps by changind the error to `sts:GetCallerIdentity: InvalidClientTokenId` which is way more common and easier to reason about.

Signed-off-by: Moulick Aggarwal <Moulick@users.noreply.github.com>